Refactor default user registration
